edit/delete only for logged in user

+++ frontend/src/Pages/ListingDetails.jsx
@@ -1,18 +1,26 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 
-const ListingDetails = (props) => {
+const ListingDetails = (props, { user }) => {
+    // console.log(props)
 
     const loaded = () => {
         const id = props.match.params.id
         const listings = props.listings
 
         const listing = listings.find(l => l._id === id)
-        console.log(listing)
-        
+        // console.log(listing)
+
         const removeListing = () => {
-            props.deleteListing(listing._id)
-            props.history.push('/shop')
+            // prompt user to confirm if they want to delete listing
+            const response = window.confirm(`Are you sure you want to delete ${listing.name}?`)
+
+            if (response) {
+                props.deleteListing(listing._id)
+                props.history.push('/shop')
+            } else {
+                return
+            }
         }
 
         return (
@@ -26,9 +34,16 @@ const ListingDetails = (props) => {
                     ${listing.price}
                 </p>
                 <p>{listing.description}</p>
-                <Link className="btn btn-primary" to={`/edit/${id}`}>Edit</Link>
-                <button className="btn btn-primary" onClick={removeListing}>Delete</button>
-            </div>
+
+                {user ?
+                    <>
+                        <Link className="btn btn-primary" to={`/edit/${id}`}>Edit</Link>
+                        <button className="btn btn-primary" onClick={removeListing}>Delete</button>
+                    </>
+                    :
+                    null
+                }
+            </div >
         )
     }
